Fix Repository Implementations & DbContext Lifecycle

+++ SampleApp/BackEnd/Infrastructure/Repository Implementations/ClientRepository.cs
@@ -2,62 +2,84 @@
 using BackEnd.Domain.Client;
 using BackEnd.Domain.Client.Site;
 using BackEnd.Infrastructure.Data;
+using Microsoft.EntityFrameworkCore;
 using System;
 using System.Linq;
 using System.Linq.Expressions;
 
 namespace BackEnd.Infrastructure.Repository_Implementations
 {
-    public class ClientRepository : IClientRepository<Client>, IDisposable
+    public class ClientRepository : IClientRepository
     {
         private readonly AppDbContext _context;
         public ClientRepository(AppDbContext context)
         {
             _context = context;
         }
-
-        public IQueryable<Client> GetClientsbyID(Guid id)
+        public async Task AddAsync(Client client)
         {
-            return _context.Clients.Where(c => c.ID == id);
+            _context.Clients.Add(client);
+            await _context.SaveChangesAsync();
+           
         }
-        public IQueryable<Client> GetClientsByName(string name)
+
+        public async Task DeleteAsync(Guid id)
         {
-            return _context.Clients.Where(c => c.Name == name);
+            var client = _context.Clients.FindAsync(id);
+            if (client != null)
+            {
+                _context.Clients.Remove(client);
+                await _context.SaveChangesAsync();
+            }
+            
+        }
+
+        public async Task UpdateAsync(Client cleint)
+        {
+            _context.Clients.Update(client);
+            await _context.SaveChangesAsync();
         }
         
-        public async Task AddAsync(Client entity)
+        
+        public async Task<List<string>> GetClientsByNameAsync(string name)
         {
-            await _context.Clients.AddAsync(entity);
-            await _context.SaveChangesAsync();
-        }
-        public async Task DeleteAsync(Client entity)
-        {
-            _context.Clients.Remove(entity);
-            await _context.SaveChangesAsync();
-        }
-        public async Task UpdateAsync(Client entity)
-        {
-            _context.Clients.Update(entity);
-            await _context.SaveChangesAsync();
-        }
-        public IQueryable<Site> GetSitesByClientID(Guid clientId)
-        {
-            return _context.Sites.Where(s => s.ClientID == clientId);
-        }
-        public async Task AddSiteAsync(Site site)
-        {
-            await _context.Sites.AddAsync(site);
-            await _context.SaveChangesAsync();
-        }
-        public async Task DeleteSiteAsync(Site site)
-        {
-            _context.Sites.Remove(site);
-            await _context.SaveChangesAsync();
+            var client =  await _context.Clients
+                .Where(c => c.Name == name)
+                .ToListAsync();
+            return client;
+            
         }
 
-        public void Dispose()
+        public async Task AddSitesAsync(Site site)
         {
-            ((IDisposable)_context).Dispose();
+            _context.Sites.Add(site);
+            await _context.SaveChangesAsync();
         }
+        public async Task DeleteSiteAsync(Guid id)
+        {
+            var site = _context.Sites.FindAsync(id)
+            if (site != null)
+            {
+                _context.Sites.Remove(site);
+                await _context.SaveChangesAsync();
+
+            }
+        }
+
+        public async Task<List<Client>> GetAllClientsAsync()
+        {
+            return await _context.Clients.ToListAsync();
+        }
+
+        public async Task<List<Site>> GetSitesByClientIdAsync(Guid clientId)
+        {
+            var sites = await _context.Sites 
+                .Where(s => s.ClientId == clientId)
+                .ToListAsync();
+            return sites;
+        }
+
+
+
     }
 }
